==> code/06_merge_bmi_csci_w_ffm_by_year.R <==
## 06 Calc BMI Metrics and Merge with Flow datasets
## R. Peek
# Calc bug metrics with CSCI and merge with flow metrics for respective datasets (annual, lag1, lag2, POR)

# Libraries ---------------------------------------------------------------

library(tidyverse)
library(sf)
library(mapview)
library(lubridate)
library(tidyverse)
library(tidylog)

# Load Data ---------------------------------------------------------------

# BMI data:

### ALL THE DATA
load(file = "data_output/02c_selected_final_bmi_dat_all.rda")
# bmi_final_dat (all BMI sites w CSCI) (full date range, Mar-Nov)
# bmi_not_selected_v2 (all BMI sites not selected bc no site pair)
# gages_not_selected_v2 (all gage sites not selected bc no site pair)
# gages_selected_v2 (all gage sites selected)
# hucs_not_selected_v2 (all hucs not selected bc no site pair)
# hucs_selected_v2 (all hucs selected)


### bmi_csci_trim (all BMI data for selected site pairs btwn May-Sep)
bmi_csci_trim <- read_rds("data_output/02c_selected_final_bmi_csci_dat_trim.rds")

### full joined FFC dataset
bmi_csci_por_trim <- read_rds("data_output/04_selected_csci_ffm_por_trim.rds")

# REVISED ECO REGIONS: unique BMI stations w ecoregions
#bmi_final_dat_ecoreg
load(file="data_output/05a_bmi_final_dat_ecoreg.rda")
ecoregs_bmi_stations <- bmi_final_dat_ecoreg %>% select(StationCode, US_L3_mod, geometry)
rm(bmi_final_dat_ecoreg)

# nhd streamlines
#load("data_output/03_selected_nhd_mainstems_gages.rda") # mainstems_all

# get Functional Flow alteration status data 
ffc_dat <- read_rds(file = url("https://github.com/ryanpeek/ffm_comparison/raw/main/output/ffc_combined/usgs_combined_alteration.rds"))

# Set Basemaps ------------------------------------------------------------

# set background basemaps:
basemapsList <- c("Esri.WorldTopoMap", "Esri.WorldImagery","Esri.NatGeoWorldMap",
                  "OpenTopoMap", "OpenStreetMap", 
                  "CartoDB.Positron", "Stamen.TopOSMFeatures")
mapviewOptions(basemaps=basemapsList)

# Make ANNUAL Dataset -----------------------------------------------------

#g_all_ffc %>% group_by(gage_id, Year) %>% distinct() %>% dim()

# get raw functional flow metrics (for every year)
ffm <- g_all_ffc %>% mutate(Year=as.integer(Year)) %>% 
  distinct() # filter out duplication

# make it long not wide for joins:
ffm <- pivot_longer(ffm, cols= c(DS_Dur_WS:Peak_Fre_5), names_to = "ffm_metric", values_to = "ffm_value") %>% 
  # drop nas
  filter(!is.na(ffm_value))

## MAKE UNTRIMMED DATA SET (ALL SAMPLE BMI MONTHS)
# untrimmed = 349
bmi_sampleid <- bmi_coms_dat %>% st_drop_geometry() %>% 
  mutate(gage_id_c = gsub("^T", "", ID)) %>% 
  dplyr::distinct(SampleID, ID, .keep_all = TRUE)

# join for annual data by GAGE ID, and BMI Sample YEAR = USGS Year
bmi_csci_ffm_ann <- left_join(bmi_sampleid, ffm, by=c("gage_id_c"="gage_id", "YYYY"="Year"))


# Make LAGGED Dataset -----------------------------------------------------

# years -1  -2
lag_yrs_1 <- unique(bmi_sampleid$YYYY) - 1
lag_yrs_2 <- unique(bmi_sampleid$YYYY) - 2

# make lag data
ffm_lag1 <- ffm %>% filter(Year %in% lag_yrs_1) %>% 
  mutate(year_flow = Year-1) # add for labeling purposes
ffm_lag2 <- ffm %>% filter(Year %in% lag_yrs_2) %>% 
  mutate(year_flow = Year-2) # add for labeling purposes

# rejoin
bmi_csci_ffm_lag1 <- left_join(bmi_sampleid, ffm_lag1, by=c("gage_id_c"="gage_id", "YYYY"="Year"))
bmi_csci_ffm_lag2 <- left_join(bmi_sampleid, ffm_lag2, by=c("gage_id_c"="gage_id", "YYYY"="Year"))

# make just sampleID dataset (trimmed = 300)
bmi_sampleid_trimmed <- bmi_coms_dat_trim %>% st_drop_geometry() %>% 
  mutate(gage_id_c = gsub("^T", "", ID)) %>% 
  dplyr::distinct(SampleID, ID, .keep_all = TRUE)

# join for annual data by GAGE ID, and BMI Sample YEAR = USGS Year
bmi_csci_ffm_ann_trim <- left_join(bmi_sampleid_trimmed, ffm, by=c("gage_id_c"="gage_id", "YYYY"="Year"))

# years -1  -2
lag_yrs_1_trim <- unique(bmi_sampleid_trimmed$YYYY) - 1
lag_yrs_2_trim <- unique(bmi_sampleid_trimmed$YYYY) - 2

# make lag data
ffm_lag1_trimmed <- ffm %>% filter(Year %in% lag_yrs_1_trim) %>% 
  mutate(year_flow = Year-1) # add for labeling purposes
ffm_lag2_trimmed <- ffm %>% filter(Year %in% lag_yrs_2_trim) %>% 
  mutate(year_flow = Year-2) # add for labeling purposes

# rejoin
bmi_csci_ffm_lag1_trim <- left_join(bmi_sampleid_trimmed, ffm_lag1_trimmed, by=c("gage_id_c"="gage_id", "YYYY"="Year"))
bmi_csci_ffm_lag2_trim <- left_join(bmi_sampleid_trimmed, ffm_lag2_trimmed, by=c("gage_id_c"="gage_id", "YYYY"="Year"))


# Add HUC Regions --------------------------------------------------

# make spatial:
bmi_csci_ffm_ann <- bmi_csci_ffm_ann %>% st_as_sf(coords=c("longitude","latitude"), crs=4326, remove=FALSE)
bmi_csci_ffm_ann_trim <- bmi_csci_ffm_ann_trim %>% st_as_sf(coords=c("longitude","latitude"), crs=4326, remove=FALSE)
bmi_csci_ffm_lag1 <- bmi_csci_ffm_lag1 %>% st_as_sf(coords=c("longitude","latitude"), crs=4326, remove=FALSE)
bmi_csci_ffm_lag1_trim <- bmi_csci_ffm_lag1_trim %>% st_as_sf(coords=c("longitude","latitude"), crs=4326, remove=FALSE)
bmi_csci_ffm_lag2 <- bmi_csci_ffm_lag2 %>% st_as_sf(coords=c("longitude","latitude"), crs=4326, remove=FALSE)
bmi_csci_ffm_lag2_trim <- bmi_csci_ffm_lag2_trim %>% st_as_sf(coords=c("longitude","latitude"), crs=4326, remove=FALSE)

# transform to match proj
ecoregs <- ecoregs %>% st_transform(4326)

# join with regions and add huc_region, make sure both df are in 4326
bmi_csci_ffm_ann <- st_join(bmi_csci_ffm_ann, left = TRUE, ecoregs["huc_region"])
bmi_csci_ffm_ann_trim <- st_join(bmi_csci_ffm_ann_trim, left = TRUE, ecoregs["huc_region"])
bmi_csci_ffm_lag1 <- st_join(bmi_csci_ffm_lag1, left = TRUE, ecoregs["huc_region"])
bmi_csci_ffm_lag1_trim <- st_join(bmi_csci_ffm_lag1_trim, left = TRUE, ecoregs["huc_region"])
bmi_csci_ffm_lag2 <- st_join(bmi_csci_ffm_lag2, left = TRUE, ecoregs["huc_region"])
bmi_csci_ffm_lag2_trim <- st_join(bmi_csci_ffm_lag2_trim, left = TRUE, ecoregs["huc_region"])

## sites to add to central valley
cvalley_add <- c("514FC1278", "514RCR001", "534DCC167")

## sites to add to great_basin
gbasin_add <- c("603MAM004", "630PS0005")

## sites to add to southcoast
scoast_add <- c("628PS1307","628PS1179","719MISSCK","719TRMDSS","719FCA001")

# Amargosa site is "609PS0053" = mojave?

# use case_when to replace: ANN
bmi_csci_ffm_ann_trim <- bmi_csci_ffm_ann_trim %>%
  mutate(huc_region = case_when(
    StationCode %in% cvalley_add ~ "central_valley",
    StationCode %in% gbasin_add ~ "great_basin",
    StationCode %in% scoast_add ~ "south_coast",
    TRUE ~ huc_region))

bmi_csci_ffm_ann <- bmi_csci_ffm_ann %>%
  mutate(huc_region = case_when(
    StationCode %in% cvalley_add ~ "central_valley",
    StationCode %in% gbasin_add ~ "great_basin",
    StationCode %in% scoast_add ~ "south_coast",
    TRUE ~ huc_region))

# use case_when to replace: LAG1
bmi_csci_ffm_lag1_trim <- bmi_csci_ffm_lag1_trim %>%
  mutate(huc_region = case_when(
    StationCode %in% cvalley_add ~ "central_valley",
    StationCode %in% gbasin_add ~ "great_basin",
    StationCode %in% scoast_add ~ "south_coast",
    TRUE ~ huc_region))

bmi_csci_ffm_lag1 <- bmi_csci_ffm_lag1 %>%
  mutate(huc_region = case_when(
    StationCode %in% cvalley_add ~ "central_valley",
    StationCode %in% gbasin_add ~ "great_basin",
    StationCode %in% scoast_add ~ "south_coast",
    TRUE ~ huc_region))

# use case_when to replace: LAG2
bmi_csci_ffm_lag2_trim <- bmi_csci_ffm_lag2_trim %>%
  mutate(huc_region = case_when(
    StationCode %in% cvalley_add ~ "central_valley",
    StationCode %in% gbasin_add ~ "great_basin",
    StationCode %in% scoast_add ~ "south_coast",
    TRUE ~ huc_region))

bmi_csci_ffm_lag2 <- bmi_csci_ffm_lag2 %>%
  mutate(huc_region = case_when(
    StationCode %in% cvalley_add ~ "central_valley",
    StationCode %in% gbasin_add ~ "great_basin",
    StationCode %in% scoast_add ~ "south_coast",
    TRUE ~ huc_region))

# SAVE OUT ----------------------------------------------------------------


save(bmi_csci_ffm_ann, file="data_output/06_selected_bmi_csci_ffm_ann.rda")
save(bmi_csci_ffm_ann_trim, file="data_output/06_selected_bmi_csci_ffm_ann_trim.rda")
save(bmi_csci_ffm_lag1, file="data_output/06_selected_bmi_csci_ffm_lag1.rda")
save(bmi_csci_ffm_lag1_trim, file="data_output/06_selected_bmi_csci_ffm_lag1_trim.rda")
save(bmi_csci_ffm_lag2, file="data_output/06_selected_bmi_csci_ffm_lag2.rda")
save(bmi_csci_ffm_lag2_trim, file="data_output/06_selected_bmi_csci_ffm_lag2_trim.rda")


# zz: Calc Bug Metrics --------------------------------------------------------

# only need to do this once, don't rerun, takes a fair bit of time
# first filter to years of interest
library(CSCI)
library(BMIMetrics)

bmi_filt <- bmi_coms_dat %>% st_drop_geometry() #%>% 
  #filter(YYYY %in% c(1993:2017))

library(purrr)

bugs_split <- bmi_filt %>%
  split(.$SampleID) %>%
  map(~BMI(.x)) # make into BMI object

bugs_samp <- bugs_split %>%
  map(~sample(.x)) # subsample to 500 individuals and aggregate

# aggregate across these splits (this takes awhile)
bugs_agg <- bugs_samp %>%
  map(~aggregate(.x))

# Calculate metrics at SAFIT Level 1
bug_metrics <- bugs_agg %>%
  map(~BMIall(.x, effort=1))

# make clean station set:
bmi_sampleids <- bmi_filt %>% distinct(SampleID, ID, .keep_all = T) %>%
  select(StationCode:HUC_12, ID, SampleID:sampledate)

# flatten and rejoin with station data:
bmi_metrics_df <- bug_metrics %>%
  do.call(what = rbind) %>%
  remove_rownames() %>%
  inner_join(., bmi_sampleids, by="SampleID")

# clean workspace, rm old bits
rm(bugs_agg, bugs_samp, bugs_split, bug_metrics, bmi_filt)

# all IDs accounted for? (should be equal to total number of rows in dataframe)
dim(bmi_metrics_df[!is.na(bmi_metrics_df$ID),])[1]

# SAVE IT
save(bmi_metrics_df, file="data_output/06_selected_bmi_csci_and_bug_metrics.rda")
